Simplify fix with runtime platform check
