+++ common/terminal/colors.go
@@ -15,7 +15,12 @@
 package terminal
 
 import (
+	"fmt"
+	"os"
+	"runtime"
 	"strings"
+
+	isatty "github.com/mattn/go-isatty"
 )
 
 const (
@@ -23,6 +28,17 @@ const (
 	warningColor = "\033[0;33m%s\033[0m"
 	noticeColor  = "\033[1;36m%s\033[0m"
 )
+
+// IsTerminal return true if the file descriptor is terminal and the TERM
+// environment variable isn't a dumb one.
+func IsTerminal(f *os.File) bool {
+	if runtime.GOOS == "windows" {
+		return false
+	}
+
+	fd := f.Fd()
+	return os.Getenv("TERM") != "dumb" && (isatty.IsTerminal(fd) || isatty.IsCygwinTerminal(fd))
+}
 
 // Notice colorizes the string in a noticeable color.
 func Notice(s string) string {
@@ -37,6 +53,16 @@ func Error(s string) string {
 // Warning colorizes the string in a colour that warns.
 func Warning(s string) string {
 	return colorize(s, warningColor)
+}
+
+// colorize s in color.
+func colorize(s, color string) string {
+	if runtime.GOOS == "windows" {
+		return s
+	}
+
+	s = fmt.Sprintf(color, doublePercent(s))
+	return singlePercent(s)
 }
 
 func doublePercent(str string) string {
