--- common/terminal/colors_others.go
@@ -1,36 +1,0 @@
-// Copyright 2018 The Hugo Authors. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// +build !windows
-
-package terminal
-
-import (
-	"fmt"
-	"os"
-
-	isatty "github.com/mattn/go-isatty"
-)
-
-// IsTerminal return true if the file descriptor is terminal and the TERM
-// environment variable isn't a dumb one.
-func IsTerminal(f *os.File) bool {
-	fd := f.Fd()
-	return os.Getenv("TERM") != "dumb" && (isatty.IsTerminal(fd) || isatty.IsCygwinTerminal(fd))
-}
-
-// colorize s in color.
-func colorize(s, color string) string {
-	s = fmt.Sprintf(color, doublePercent(s))
-	return singlePercent(s)
-}
